Add files via upload
This PHP script establishes a connection to a MySQL database using the mysqli_connect function. It specifies the database host, user, password, and database name as connection parameters:

Database Host: localhost
Database User: id22229389_root
Database Password: UPcs2024.
Database Name: id22229389_databaza
The script includes error handling to exit and display an error message if the connection fails. If the connection is successful, the script proceeds without displaying any messages. If the connection fails, it exits and prints an error message with the specific reason for the failure.